WorkDescription type + MAX_SIMILAR_CONCURRENT_WORK const
Move work_pool into own module
