+++ src/server/dispatch.rs
@@ -9,6 +9,8 @@
 // except according to those terms.
 
 use super::requests::*;
+use actions::work_pool;
+use actions::work_pool::WorkDescription;
 use jsonrpc_core as jsonrpc;
 use server;
 use server::{Request, Response};
@@ -48,7 +50,7 @@ macro_rules! define_dispatch_request_enum {
                         let Request { id, params, received, .. } = req;
                         let timeout = $request_type::timeout();
 
-                        let receiver = receive_from_thread(move || {
+                        let receiver = work_pool::receive_from_thread(move || {
                             // checking timeout here can prevent starting expensive work that has
                             // already timed out due to previous long running requests
                             // Note: done here on the threadpool as pool scheduling may incur
@@ -59,7 +61,7 @@ macro_rules! define_dispatch_request_enum {
                             else {
                                 $request_type::handle(ctx, params)
                             }
-                        }, $request_type::METHOD);
+                        }, WorkDescription($request_type::METHOD));
 
                         match receiver.recv_timeout(timeout)
                             .unwrap_or_else(|_| $request_type::fallback_response()) {
